added the pgp plugin; initial draft
