+++ buildSrc/src/main/resources/META-INF/gradle-plugins/checksumverify.properties
@@ -1,0 +1,1 @@
+implementation-class=ChecksumPlugin
