--- buildSrc/src/main/resources/META-INF/gradle-plugins/witness.properties
@@ -1,1 +1,0 @@
-implementation-class=WitnessPlugin
