+++ gradle/pgpverify.gradle
@@ -1,3 +1,3 @@
-pgpPlugin {
-    enabled = Boolean.getBoolean("verifyPGPs")
+PGPVerifyPlugin {
+    verifyPomFiles = true
 }
